Refactoring tabs for creation PO page
